requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
Jinja2==3.1.6
asyncio==3.4.3
python-dotenv==1.1.0
pydantic==2.11.3

